# Improve `CowData::insert` performance