@aaronp64 2024-07-15
Replacing the first two `size()` calls by just calling it once up front looks pretty straightforward, I'm not sure what to do with the third `size()` call though (for the `for` loop).  I think if we use a value from before `resize` is called, it might technically change the behavior in the case where `resize` fails - if a `realloc` was required and failed, `_ptr` might point to memory smaller than the requested size.  Without checking `size()` again, we might go past the end of the array if we assume `resize` was successful.  Not positive I'm following the allocation/error logic correctly though, might not be an issue.

Should I leave the `size()` call as is in the `for` loop, or replace all of them?