@RandomShaper 2024-07-15
I'd go for all of them. And if you want to handle an error from `resize()`, feel free. The current implementation is not doing it (just assuming the resize will never fail) and, although it's not that important because you're actually doomed if that fails, there's at least a theoretical correctness point in using it.

I'd go for something like this:
```C++
	Error insert(Size p_pos, const T &p_val) {
		Size curr_size = size();
		ERR_FAIL_INDEX_V(p_pos, curr_size + 1, ERR_INVALID_PARAMETER);
		Error err = resize(curr_size + 1);
		ERR_FAIL_COND_V(err, err);
		T *p = ptrw();
		for (Size i = curr_size; i > p_pos; i--) {
			p[i] = p[i - 1];
		}
		p[p_pos] = p_val;

		return OK;
	}
```