@aaronp64 2024-07-15
Updated and added error check