@MewPurPur 2024-07-20
For a user who doesn't know about cow data, what will be sped up by this optimization?