@AThousandShips 2024-07-20
This will improve things like `Vector<T>` insert (including all packed arrays)