@MewPurPur 2024-07-20
Presumably this will affect things all over the engine where Vector<T>.insert() is used?