@AThousandShips 2024-07-20
Might not be significant given other performance considerations, but it should affect them at least somewhat